+++ examples/data-sources/voipms_callback/data-source.tf
@@ -1,3 +1,3 @@
 data "voipms_callback" "mobile" {
-  id = "3001"
+  description = "Mobile"
 }