+++ examples/data-sources/voipms_phonebook_group/data-source.tf
@@ -1,3 +1,3 @@
 data "voipms_phonebook_group" "spam" {
-  id = "5001"
+  name = "Spam"
 }