+++ examples/resources/voipms_did/resource.tf
@@ -1,13 +1,31 @@
+locals {
+  my_pops = {
+    ny7 = "newyork7.voip.ms"
+  }
+}
+
+data "voipms_voicemail" "johns_voicemail" {
+  name = "John"
+}
+
+data "voipms_forwarding" "mobile" {
+  description = "Mobile"
+}
+
+data "voipms_subaccount" "gateway" {
+  username = "gateway"
+}
+
 resource "voipms_did" "home" {
   did                  = "5550001001"
   note                 = "Home line"
-  routing              = "account:100001_gateway"
-  pop                  = 73
+  routing              = "account:${data.voipms_subaccount.gateway.account}"
+  pop_hostname         = local.my_pops.ny7
   dialtime             = 30
-  voicemail            = "101"
-  failover_busy        = "vm:101"
-  failover_noanswer    = "vm:101"
-  failover_unreachable = "vm:101"
+  voicemail            = data.voipms_voicemail.johns_voicemail.mailbox
+  failover_busy        = "vm:${data.voipms_voicemail.johns_voicemail.mailbox}"
+  failover_noanswer    = "vm:${data.voipms_voicemail.johns_voicemail.mailbox}"
+  failover_unreachable = "vm:${data.voipms_voicemail.johns_voicemail.mailbox}"
 
   sms_enabled     = true
   webhook         = "https://example.lambda-url.us-east-1.on.aws/"